refactor: add attribute platform
