--- src/app/services/types.ts
@@ -1,4 +1,0 @@
-export const enum PLATFORM {
-  YOUTUBE = 'youtube',
-}
-
