+++ src/app/utils/getPlatform.ts
@@ -1,0 +1,4 @@
+export default (): string => {
+  const app = document.querySelector('#op-app') as HTMLElement
+  return app.getAttribute('platform') as string
+}
